requirements.txt
altair==4.0.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dlib==19.24.6
entrypoints==0.4
face-recognition==1.3.0
face_recognition_models==0.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.26.0
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==3.20.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
semver==3.0.4
six==1.17.0
smmap==5.0.2
streamlit==1.12.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
validators==0.34.0
watchdog==6.0.0
zipp==3.21.0
